requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
python-multipart==0.0.9
